requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
distro==1.9.0
dotenv==0.9.9
filelock==3.24.3
filetype==1.2.0
fire==0.7.1
fonttools==4.61.1
fsspec==2026.2.0
google-auth==2.48.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.7
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
mss==10.1.0
networkx==3.6.1
numpy==2.4.2
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
packaging==26.0
pi_heif==1.2.1
pillow==12.1.1
pillow-avif-plugin==1.5.5
polars==1.38.1
polars-runtime-32==1.38.1
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybboxes==0.1.6
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.14
rsa==4.9.1
sahi==0.11.36
scipy==1.17.0
setuptools==82.0.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.4
termcolor==3.3.0
terminaltables==3.1.10
torch==2.10.0+cu130
torchvision==0.25.0+cu130
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.14
ultralytics-thop==2.0.18
urllib3==2.6.3
websockets==15.0.1
